requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
numpy>=1.19.0
scikit-learn>=1.0.0
nltk>=3.6.0
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.2
keybert>=0.5.0
spacy>=3.0.0
librosa>=0.8.0
soundfile>=0.10.0
thinc>=8.0.0
tensorflow-cpu>=2.12.0
